[Refactor] Swiper 컴포넌트 사용과 뷰, 데이터 계층 분리
